+++ healpy/projaxes.py
@@ -20,7 +20,7 @@
 import logging
 
 log = logging.getLogger("healpy")
-from astropy.utils.decorators import deprecated_renamed_argument
+from .utils.deprecation import deprecated_renamed_argument
 from . import projector as P
 from . import rotator as R
 from . import pixelfunc